Fix Generate wrong file path

+++ app/File.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 use \File as FS;
 class File extends Model
 {
-	// Do not guard anything
+    // Do not guard anything
     protected $guarded = [];
 
 
@@ -25,10 +25,10 @@ class File extends Model
      * @return boolean|string URL to file or null if file was deled
      */
     public function url(){
-    	$realPath = public_path($this->file_path);
-    	if (file_exists($realPath))
-    		return asset($this->file_path);
-    	return false;
+        $realPath = public_path($this->file_path);
+        if (file_exists($realPath))
+            return asset($this->file_path);
+        return false;
     }
 
     
@@ -37,6 +37,8 @@ class File extends Model
      * @return boolean Success?
      */
     public function remove() {
-		return FS::deleteDirectory(public_path(dirname($this->file_path)));
+        if (file_exists(public_path(dirname($this->file_path))))
+          return FS::deleteDirectory(public_path(dirname($this->file_path)));
+        return false;
     }
 }
